imgmemes.py: fixing up the .tweet code

+++ userbot/modules/imgmemes.py
@@ -306,12 +306,14 @@ async def tweet(event):
     else:
         await event.edit("`What should i tweet? Give your username and tweet!`"
                          )
-    await event.edit(f"`Requesting {username} to tweet...`")
-    text = deEmojify(text)
-    img = await tweets(text, username)
-    await event.client.send_file(event.chat_id, img, reply_to=reply_to_id)
-    await event.delete()
-    await purge()
+    if 'username' not in locals():
+        await event.edit(f"`Send a username to tweet\n like .tweet username.Text to tweet...`")
+    else:  
+        text = deEmojify(text)
+        img = await tweets(text, username)
+        await event.client.send_file(event.chat_id, img, reply_to=reply_to_id)
+        await event.delete()
+        await purge()
 
 
 @register(pattern="^.threat(?: |$)(.*)", outgoing=True)
